requirements.txt
about-time==4.2.1
alive-progress==3.1.5
certifi==2024.7.4
charset-normalizer==2.0.12
click==8.1.7
Flask==2.0.1
grapheme==0.6.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
pydub==0.25.1
python-dotenv==0.21.1
requests==2.26.0
urllib3==1.26.19
Werkzeug==2.0.1
gunicorn==22.0.0
pyngrok==7.0.5
openai==1.33.0